Stop overwriting unprocessed VEP VCF with processed one
The human VEP output VCF was found to contain entries that
had not been processed by the process_output analysis, despite no
failed process_output jobs or errors in the logs explaining this.
Instead of overwriting the unprocessed VEP VCF file with the
processed one the pipeline now just deletes the unprocessed one
and combines the files with the '.processed' suffix in the
combine_output step.  As such it won't be possible for the final
output to include unprocessed entries.  Resolves AGR-2985.